Refactor stroke style management to reuse code
